+++ scripts/helpers/paths.js
@@ -37,6 +37,7 @@ module.exports = {
     node_modules,
     '@salesforce-ux/design-system-primitive-tokens'
   ),
+  sldsStylingHooksProps: rootPath('styling-hooks'),
   sdsStylingHooksProps: path.resolve(
     node_modules,
     '@salesforce-ux/sds-styling-hooks/src/props'