+++ ui/components/popovers/RELEASENOTES.md
@@ -1,6 +1,13 @@
 <!-- Release notes authoring guidelines: http://keepachangelog.com/ -->
 
 # Popovers Release Notes
+## 2.21.0
+
+### Added button to the story `body-max-height-small` to fix scrollable-region-focusable wcag issue.
+
+### Added
+
+- Added `title` attribute to meet accessibility requirements.
 
 ## 2.19